Moving the check for node version to a sdk file.
